[FPGA][NFC] improves tests for no-global-work-offset attribute
